// ui/src/lib/api/model/asynqServerInfo.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform using task queues for scalable browser session management.
 * OpenAPI spec version: 2.0
 */
import type { AsynqWorkerInfo } from './asynqWorkerInfo';
import type { AsynqServerInfoQueues } from './asynqServerInfoQueues';

export interface AsynqServerInfo {
  /** A List of active workers currently processing tasks. */
  activeWorkers?: AsynqWorkerInfo[];
  /** Server configuration details.
See Config doc for field descriptions. */
  concurrency?: number;
  /** Host machine on which the server is running. */
  host?: string;
  /** Unique Identifier for the server. */
  id?: string;
  /** PID of the process in which the server is running. */
  pid?: number;
  queues?: AsynqServerInfoQueues;
  /** Time the server started. */
  started?: string;
  /** Status indicates the status of the server.
TODO: Update comment with more details. */
  status?: string;
  strictPriority?: boolean;
}